requirements.txt
pyyaml==3.12
psycopg2==2.7.3.2
Django==2.0
djangorestframework==3.7.7
django-rest-auth==0.9.2
djangorestframework-jwt==1.11.0
django-allauth==0.34.0
django-mptt==0.9.0
django-taggit==0.22.2
django-taggit-serializer==0.1.5
cryptography==2.1.4